Fix: Ensure proper app initialization after permission changes
Reason:
Previously, when system permissions (e.g., CAMERA) were modified, Android restarted the DashboardActivity without reinitializing critical components like SharedPreferences and the database. This led to app crashes and unexpected behavior.

Solution:
- Moved global initialization ( preferences) to the Application class to ensure they persist across activity restarts.
- Added an unlock database in DashboardActivity to verify initialization before proceeding. If initialization fails, the user is redirected to MainActivity.
- Ensured that essential components are properly reloaded when the app resumes after permission changes.

Now, the app remains stable even after system permission modifications.
