+++ presentation/src/main/AndroidManifest.xml
@@ -73,6 +73,7 @@
         <activity
             android:name=".view.screens.main.MainActivity"
             android:exported="true"
+            android:launchMode="singleTask"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
